+++ crates/ryo3-twox-hash/src/xxhash32.rs
@@ -3,16 +3,15 @@ use pyo3::intern;
 use pyo3::prelude::*;
 use pyo3::types::{PyModule, PyModuleMethods, PyString};
 use pyo3::{Bound, PyResult, Python, pyfunction, wrap_pyfunction};
-use ryo3_core::PyLock;
+use ryo3_core::PyMutex;
 use std::hash::Hasher;
-use std::sync::Mutex;
 use twox_hash::XxHash32;
 
 #[pyclass(name = "xxh32", frozen)]
 #[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3.xxhash"))]
 pub struct PyXxHash32 {
     seed: u32,
-    hasher: Mutex<XxHash32>,
+    hasher: PyMutex<XxHash32, true>,
 }
 
 #[pymethods]
@@ -28,12 +27,12 @@ impl PyXxHash32 {
                 hasher.write(s.as_ref());
                 Self {
                     seed,
-                    hasher: Mutex::new(hasher),
+                    hasher: hasher.into(),
                 }
             }
             None => Self {
                 seed,
-                hasher: Mutex::new(hasher),
+                hasher: hasher.into(),
             },
         }
     }
@@ -80,17 +79,18 @@ impl PyXxHash32 {
     }
 
     #[expect(clippy::needless_pass_by_value)]
-    fn update(&self, data: ryo3_bytes::PyBytes) -> PyResult<()> {
-        // self.hasher.update(b.as_ref());
-        let mut hasher = self.hasher.py_lock()?;
-        hasher.write(data.as_ref());
-        Ok(())
+    fn update(&self, py: Python<'_>, data: ryo3_bytes::PyBytes) -> PyResult<()> {
+        py.detach(|| {
+            let mut hasher = self.hasher.py_lock()?;
+            hasher.write(data.as_ref());
+            Ok(())
+        })
     }
 
     fn copy(&self) -> PyResult<Self> {
         let hasher = self.hasher.py_lock()?;
         Ok(Self {
-            hasher: Mutex::new(hasher.clone()),
+            hasher: hasher.clone().into(),
             seed: self.seed,
         })
     }
@@ -104,8 +104,8 @@ impl PyXxHash32 {
     #[expect(clippy::needless_pass_by_value)]
     #[staticmethod]
     #[pyo3(signature = (data, *, seed = None))]
-    fn oneshot(data: ryo3_bytes::PyBytes, seed: Option<u32>) -> u32 {
-        XxHash32::oneshot(seed.unwrap_or(0), data.as_ref())
+    fn oneshot(py: Python<'_>, data: ryo3_bytes::PyBytes, seed: Option<u32>) -> u32 {
+        py.detach(|| XxHash32::oneshot(seed.unwrap_or(0), data.as_ref()))
     }
 }
 
@@ -115,23 +115,26 @@ impl PyXxHash32 {
 #[expect(clippy::needless_pass_by_value)]
 #[pyfunction]
 #[pyo3(signature = (data, *, seed = None))]
-pub fn xxh32_digest(data: ryo3_bytes::PyBytes, seed: Option<u32>) -> PyDigest<u32> {
-    let digest = twox_hash::XxHash32::oneshot(seed.unwrap_or(0), data.as_ref());
-    PyDigest(digest)
+pub fn xxh32_digest(py: Python<'_>, data: ryo3_bytes::PyBytes, seed: Option<u32>) -> PyDigest<u32> {
+    py.detach(|| twox_hash::XxHash32::oneshot(seed.unwrap_or(0), data.as_ref()).into())
 }
 
 #[expect(clippy::needless_pass_by_value)]
 #[pyfunction]
 #[pyo3(signature = (data, *, seed = None))]
-pub fn xxh32_intdigest(data: ryo3_bytes::PyBytes, seed: Option<u32>) -> u32 {
-    twox_hash::XxHash32::oneshot(seed.unwrap_or(0), data.as_ref())
+pub fn xxh32_intdigest(py: Python<'_>, data: ryo3_bytes::PyBytes, seed: Option<u32>) -> u32 {
+    py.detach(|| twox_hash::XxHash32::oneshot(seed.unwrap_or(0), data.as_ref()))
 }
 
 #[expect(clippy::needless_pass_by_value)]
 #[pyfunction]
 #[pyo3(signature = (data, *, seed = None))]
-pub fn xxh32_hexdigest(data: ryo3_bytes::PyBytes, seed: Option<u32>) -> PyHexDigest<u32> {
-    twox_hash::XxHash32::oneshot(seed.unwrap_or(0), data.as_ref()).into()
+pub fn xxh32_hexdigest(
+    py: Python<'_>,
+    data: ryo3_bytes::PyBytes,
+    seed: Option<u32>,
+) -> PyHexDigest<u32> {
+    py.detach(|| twox_hash::XxHash32::oneshot(seed.unwrap_or(0), data.as_ref()).into())
 }
 
 pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {