+++ crates/ryo3-pydantic/src/lib.rs
@@ -42,10 +42,12 @@ pub mod interns {
     }
 
     unpaid_intern!(_pydantic_validate);
+    unpaid_intern!(_pydantic_serialize);
     unpaid_intern!(_pydantic_validate_strict);
     unpaid_intern!(json_unless_none, "json-unless-none");
     unpaid_intern!(no_info_wrap_validator_function);
     unpaid_intern!(no_info_plain_validator_function);
+    unpaid_intern!(plain_serializer_function_ser_schema);
     unpaid_intern!(to_string_ser_schema);
     unpaid_intern!(lax_or_strict_schema);
     unpaid_intern!(when_used);
@@ -54,6 +56,7 @@ pub mod interns {
     unpaid_intern!(min_length);
     unpaid_intern!(max_length);
     unpaid_intern!(pattern);
+    unpaid_intern!(info_arg);
     // schemas
     unpaid_intern!(str_schema);
     unpaid_intern!(bytes_schema);